% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/define_arrows_layout.R
\name{define_arrow_layout}
\alias{define_arrow_layout}
\title{Create arrow connections}
\usage{
define_arrow_layout(prms = NULL, rse = NULL, advan = NULL,
  des_info = NULL, scaling = TRUE, scale.fun = function(x) {     x^(1/3)
  }, font = "Avenir", clearance = TRUE, arrow.fontsize = 12, ...)
}
\arguments{
\item{prms}{a numerical vector of parameter values}

\item{rse}{a numerical vector of parameter uncertainty}

\item{advan}{the value of the nonmem subroutine ADVAN}

\item{scaling}{logical, if \code{TRUE} arrow width and colors will be scaled.
If \code{FALSE} standard model diagram will be created}

\item{scale.fun}{function to be used for arrow width scaling}

\item{font}{font name of arrow labels}

\item{clearance}{logical, if \code{TRUE} arrows will be scaled to their available
clearance value. If \code{FALSE} arrows will be scaled to their available
rate constant value}

\item{arrow.fontsize}{font size of the arrow labels in point size}
}
\value{
A \code{data.frame}
}
\description{
Combine information for each arrow's origin, destination, label,
witdh, color, scaling and more into a dataframe
}
\details{
The default \code{scaling.fun} argument is set to \code{cubic}, to match
 each the method used on volumes. Arrow will be sized to the cubic root of the
 clearance or rate constant value.
}
\seealso{
\code{\link{import_qmd_info}}, \code{\link{qmd}}
}

