% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/define_comp_layout.R
\name{define_comp_layout}
\alias{define_comp_layout}
\title{Create compartment layout}
\usage{
define_comp_layout(qmd_info = NULL, scaling = TRUE,
  scale.fun = function(x) {     x^(1/3) }, box.ratio = 3/4, filled = TRUE,
  font = "Avenir", comp.fontsize = 12, ...)
}
\arguments{
\item{qmd_info}{a \code{qmd_info} object generated with \code{import_qmd_info} or
\code{skeleton_qmd_info}}

\item{scaling}{logical if \code{TRUE} compartment size and colors will be scaled.
If \code{FALSE} standard model diagram will be created}

\item{scale.fun}{a function to be used for compartment size scaling}

\item{box.ratio}{value for the compartment aspect ratio}

\item{filled}{logical if \code{TRUE} compartment will be filled.
If \code{FALSE} only the compartment edges will be drawn}

\item{font}{font name of the compartment labels}

\item{comp.fontsize}{font size of the compartment labels in point size}
}
\value{
A \code{data.frame}
}
\description{
Combine information for each compartment's rank, label, size,
color, scaling, etc.
}
\details{
The default \code{scaling.fun} argument is set to be the cubic root, each
 compartment will be sized to the cubic root of the volume value. This scaling
 allows for a more realistic size comparison and avoids too big differences in size.
}
\examples{
\dontrun{
qmd_info <- import_qmd_info(dir = '../models/pk/', runno = '001')
comp     <- define_comp_layout(qmd_info)
}
}
\seealso{
\code{\link{define_arrow_layout}}, \code{\link{import_qmd_info}}, \code{\link{qmd}}
}

