% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/define_graph.R
\name{define_graph}
\alias{define_graph}
\title{Create a graph object using data frames representative of comp and arrow}
\usage{
define_graph(comp = NULL, arrow = NULL, pbpk = NULL, graph_attrs = NULL,
  comp_attrs = NULL, arrow_attrs = NULL, directed = TRUE,
  graph_name = NULL, graph_time = NULL, graph_tz = NULL)
}
\arguments{
\item{comp}{a data.frame created by \code{define_comp_layout}}

\item{arrow}{a data.frame created by \code{define_arrow_layout}}

\item{pbpk}{a list created by \code{define_pbpk_layout}}

\item{graph_attrs}{an optional vector of graph attribute statements that can
serve as defaults for the graph.}

\item{comp_attrs}{an optional vector of comp attribute statements that can
serve as defaults for comp.}

\item{arrow_attrs}{an optional vector of comp attribute statements that can
serve as defaults for arrow.}

\item{directed}{logical, if \code{TRUE} (default) directed arrow operations
will be generated.}

\item{graph_name}{an optional string for labeling the graph object.}

\item{graph_time}{a date or date-time string (required for insertion of graph
into a graph series of the type temporal).}

\item{graph_tz}{an optional value for the time zone (tz) corresponding to the
date or date-time string supplied as a value to graph_time. If no time zone is
provided then it will be set to GMT.}
}
\description{
Generates a graph object using data frames for comp and arrow;
the graph object can be manipulated by other functions.
}
\examples{
\dontrun{
graph <- define_graph(comp  = define_comp_layout(examples$onecomp),
                      arrow = define_arrow_layout(examples$onecomp),
                      graph_attrs = c('splines = true',
                                      'ranksep = 0',
                                      'nodesep = 0.15',
                                      'rankdir = LR'))
DiagrammeR::render_graph(graph)
}
}

