% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_pbpk_layout.R
\name{define_pbpk_layout}
\alias{define_pbpk_layout}
\title{Make proper PBPK layout}
\usage{
define_pbpk_layout(comp = NULL, arrow = NULL, pbpk_color = TRUE,
  vein_comp_label = "venous", artery_comp_label = "arterial")
}
\arguments{
\item{comp}{a data.frame created by \code{define_comp_layout}}

\item{arrow}{a data.frame created by \code{define_arrow_layout}}

\item{pbpk_color}{logical if \code{TRUE} oxygenated blood is represented in red
and deoxygenated blood is represented in blue. If \code{FALSE} previously defined
color code in \code{define_comp_layout} and \code{define_arrow_layout} will be used}

\item{vein_comp_label}{label of the veinous compartment}

\item{artery_comp_label}{label of the arterial compartment}
}
\value{
A \code{list}
}
\description{
Make special layout for PBPK models with the veinous compartment on
the left, the arterial compartment on the right and the organs in between.
}
\examples{
\dontrun{
qmd_info <- import_qmd_info(dir = '../models/pk/', runno = '001')
comp     <- define_comp_layout(qmd_info)
arrow    <- define_arrow_layout(qmd_info)
pbpk     <- define_pbpk_layout(comp = comp, arrow = arrow)
}
}
\seealso{
\code{\link{define_comp_layout}}, \code{\link{define_arrow_layout}}, \code{\link{qmd}}
}
