% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/examples.R
\name{examples}
\alias{examples}
\title{QMD Examples}
\format{A list of 3 levels: \code{onecomp}, \code{twocomp} and \code{threecomp}}
\source{
\href{http://www.ncbi.nlm.nih.gov/pubmed/18751690}{onecomp}: D. Elsherbiny et al.
Population pharmacokinetics of nevirapine in combination with rifampicin-based short
course chemotherapy in HIV- and tuberculosis-infected South African patients.
Eur J Clin Pharmacol. 65:71–80. (2009)

\href{http://www.ncbi.nlm.nih.gov/pubmed/18519729}{twocomp}: T. Dorlo et al.
Pharmacokinetics of miltefosine in Old World cutaneous leishmaniasis patients.
Antimicrob. Agents Chemother. 52:8, 2855–2860. (2008)

\href{http://www.ncbi.nlm.nih.gov/pubmed/17662086}{threecomp}: S. Fanta et al.
Developmental pharmacokinetics of ciclosporin – a population pharmacokinetic study
in paediatric renal transplant candidates. British Journal of Clinical Pharmacology.
64:6, 772–784. (2007)
}
\description{
A list of 3 \code{qmd_info} examples on the pharmacokinetic models for
nevirapine (\code{onecomp}), miltefosine (\code{twocomp}) and
ciclosporin (\code{threecomp}).
}
\details{
The \code{examples} file contains the following:
\itemize{
\item \code{examples$onecomp} nevirapine PK model \code{qmd_info}
\item \code{examples$twocomp} miltefosine PK model \code{qmd_info}
\item \code{examples$threecomp} ciclosporin PK model \code{qmd_info}
}

Each of these three examples contain the following:
\itemize{
  \item tvprm. the model parameters typical values eg. KA (h-1), CL (L/h) and V (L)
  \item rse. the model parameters uncertainty as relative standard error (%)
  \item data. the individual model parameter read from a patab file
  \item advan. the nonmem subroutine ADVAN defining the structural model
  \item trans. the nonmem subroutine TRANS defining the parametrization
  \item des_info. the parsed differencial equations
}
}
\examples{
# One-compartment model
qmd(examples$onecomp)

# Two-compartment model
qmd(examples$twocomp)

# Three-compartment model
qmd(examples$threecomp)
}

