% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_color.R
\name{hex_color}
\alias{hex_color}
\title{Create color in hexadecimal}
\usage{
hex_color(color = NULL, alpha = NULL)
}
\arguments{
\item{color}{a hexadecimal or named color e.g. '#1874CD' or 'dodgerblue3'}

\item{alpha}{transparency factor. Numeric value between 0 and 1}
}
\description{
Create hexadecimal type color name with transparency
}
\examples{
\dontrun{
color <- hex_color(color = 'dodgerblue3', alpha = 0.5)
}
}

