% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import_qmd_info.R
\name{import_qmd_info}
\alias{import_qmd_info}
\title{Import model parameters}
\usage{
import_qmd_info(dir = NULL, prefix = "run", runno, ext = ".mod",
  file = NULL)
}
\arguments{
\item{dir}{location of the model files}

\item{prefix}{prefix of the model file name}

\item{runno}{run number to be evaluated}

\item{ext}{model file extention}
}
\value{
A list containing the parameters and RSE as \code{"data.frame"},
        the nonmem ADVAN (advan) and TRANS (trans) as numeric value.
}
\description{
Import model parameters from NONMEM run
}
\examples{
\dontrun{
prm_list <- import_qmd_info(dir='../models/pk/', runno='001')
}
}
\seealso{
\code{\link{format_qmd_info}}, \code{\link{qmd}}
}

