% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_arrows.R
\name{make_arrows}
\alias{make_arrows}
\title{Create arrow data}
\usage{
make_arrows(prms = NULL, rse = NULL, advan = NULL, scaling = TRUE,
  scale.fun = "cubic", font = "Avenir", clearance = TRUE,
  edge.fontsize = 12, ...)
}
\arguments{
\item{prms}{a numerical vector of parameter values}

\item{rse}{a numerical vector of parameter uncertainty}

\item{advan}{the nonmem $SUB ADVAN}

\item{scaling}{logical if \code{TRUE} arrow width and colors will be scaled}

\item{scale.fun}{function to be used for arrow width scaling}

\item{font}{font of arrow labels}

\item{clearance}{logical if \code{TRUE} arrows will be scaled to their available
clearance value. If \code{FALSE} arrows will be scaled to their available
rate constant value}

\item{node.fontsize}{font size of the arrow labels}
}
\value{
A \code{data.frame} of arrows
}
\description{
Creates information on arrows label, witdh, color, scaling, etc.
}
\seealso{
\code{\link{prm_import}}, \code{\link{modelviz}}
}

