% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_comp.R
\name{make_comp}
\alias{make_comp}
\title{Create compartment data}
\usage{
make_comp(prms = NA, rse = NA, advan = NULL, scaling = TRUE,
  scale.fun = "cubic", box.ratio = 3/4, filled = TRUE, font = "Avenir",
  node.fontsize = 12, ...)
}
\arguments{
\item{prms}{a numerical vector of parameter values}

\item{rse}{a numerical vector of parameter uncertainty}

\item{advan}{the nonmem $SUB ADVAN}

\item{scaling}{logical if \code{TRUE} compartment size and colors will be scaled}

\item{scale.fun}{function to be used for compartment size scaling}

\item{box.ratio}{compartment aspect ratio}

\item{filled}{logical if \code{TRUE} compartment will be filled}

\item{font}{font of the compartment labels}

\item{node.fontsize}{font size of the compartment labels}
}
\value{
A \code{data.frame} of compartments
}
\description{
Creates information on compartments label, size, color, scaling, etc.
}
\seealso{
\code{\link{prm_import}}, \code{\link{modelviz}}
}

