% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_graph.R
\name{make_graph}
\alias{make_graph}
\title{DiagrammeR functions}
\usage{
make_graph(nodes_df = NULL, edges_df = NULL, graph_attrs = NULL,
  node_attrs = NULL, edge_attrs = NULL, directed = TRUE,
  graph_name = NULL, graph_time = NULL, graph_tz = NULL)
}
\description{
Derived from DiagrammeR
}

