% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelviz.R
\name{modelviz}
\alias{modelviz}
\title{Create and render QMD}
\usage{
modelviz(data, horizontal = TRUE, ...)
}
\arguments{
\item{data}{a \code{list} containing the parameters, their RSE and the
nonmem subroutine ADVAN value}

\item{horizontal}{logical if \code{TRUE} the layout will be horizontal}

\item{...}{other arguments passed to \code{\link{make_comp}}
and \code{\link{make_arrow}} see details below for
a list of the available options}
}
\value{
A graphic object
}
\description{
Create and renders Quantitative Model Diagrams (QMD)
for NONMEM models
}
\details{
\itemize{
 \item{scaling}
 \item{scale.fun}
 \item{font}
 \item{comp.fontsize}
 \item{arrow.fontsize}
 \item{box.ratio}
 \item{filled}
 \item{clearance}
}
}
\examples{
\dontrun{
modelviz(parameters, horizontal = FALSE)
}
}
\seealso{
\code{\link{prm_import}}, \code{\link{prm_format}}
}

