% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/onecomp.R
\name{onecomp}
\alias{onecomp}
\title{Nevirapine pharmacokinetic model parameters}
\format{A list of 3 levels: prm, rse and advan.}
\source{
D. Elsherbiny et al. Population pharmacokinetics of nevirapine
in combination with rifampicin-based short course chemotherapy in
HIV- and tuberculosis-infected South African patients. Eur J Clin Pharmacol. 65:71–80. (2009)
\url{http://www.ncbi.nlm.nih.gov/pubmed/18751690}
}
\description{
A dataset containing the pharmacokinetic typical parameters values and uncertainty
for nevirapine in combination with rifampicin-based short
course chemotherapy in HIV- and tuberculosis-infected South African patients.
}
\details{
\itemize{
  \item prm. the model parameters typical values: KA (h-1), CL (L/h) and V (L)
  \item rse. the model parameters uncertainty as relative standard error (rse)
  \item advan. the nonmem subroutine ADVAN defining a 1-compartment model with
  first order absorption
}
}

