% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_des_block.R
\name{parse_des_block}
\alias{parse_des_block}
\title{Parse $DES}
\usage{
parse_des_block(des_block = NULL)
}
\arguments{
\item{des_block}{string containing the $DES block from NONMEM}
}
\value{
A \code{list}
}
\description{
Parse a $DES block from NONMEM
}
\seealso{
\code{\link{import_qmd_info}}, \code{\link{qmd}}
}

