% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nonmem_model.R
\name{parse_nonmem_model}
\alias{parse_nonmem_model}
\title{NONMEM model file parser}
\usage{
parse_nonmem_model(dir = NULL, runno = NULL, prefix = "run",
  ext = ".mod", file = NULL)
}
\arguments{
\item{dir}{location of the model files}

\item{runno}{run number to be evaluated}

\item{prefix}{prefix of the model file name}

\item{ext}{model file extention}

\item{file}{full file name as an alternative to \code{dir}, \code{prefix},
\code{runno} and \code{ext}}
}
\value{
A \code{data.frame} containing the parsed model code (\code{CODE}) as well as a numeric
(\code{LEVEL}) and character (\code{SUB}) indexing for subroutines.
}
\description{
Parse NONMEM model files in R format
}
\examples{
\dontrun{
mod_file <- parse_nonmem_model(dir = '../models/pk/', runno = '101')
}
}
\seealso{
\code{\link{import_qmd_info}}, \code{\link{read_nmtab}}
}

