% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prm_format.R
\name{prm_format}
\alias{prm_format}
\title{Format parameter vector}
\usage{
prm_format(data, advan, trans)
}
\arguments{
\item{data}{a \code{data.frame} of parameters}

\item{advan}{the nonmem $SUB ADVAN}

\item{trans}{the nonmem $SUB TRANS}
}
\value{
A \code{data.frame} of parameters
}
\description{
Translate clearances (CL, Q) and volumes (V) into
             rate constant and the other way around.
}
\examples{
\dontrun{
prm_list   <- prm_import(dir='../models/pk/', runno='001')
parameters <- prm_format(data=prm_list$prm, advan=prm_list$advan, trans=prm_list$trans)
}
}
\seealso{
\code{\link{prm_import}}, \code{\link{modelviz}}
}

