% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prm_import.R
\name{prm_import}
\alias{prm_import}
\title{Import model parameters}
\usage{
prm_import(dir = NULL, prefix = "run", runno, ext = ".mod")
}
\arguments{
\item{dir}{where are the model files located}

\item{prefix}{start of the model file name}

\item{runno}{run number of the table files to read}

\item{ext}{model file extention}
}
\value{
A list containing the parameters and RSE as \code{"data.frame"},
        the nonmem ADVAN (advan) and TRANS (trans) as numeric value.
}
\description{
Import model parameters from NONMEM run
}
\examples{
\dontrun{
prm_list <- prm_import(dir='../models/pk/', runno='001')
}
}
\seealso{
\code{\link{prm_format}}, \code{\link{modelviz}}
}

