% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd.R
\name{qmd}
\alias{qmd}
\title{Automated creation and rendering of QMD graphs}
\usage{
qmd(qmd_info = NULL, scaling = TRUE, comp_scale_fun = function(x) {    
  sqrt(x) }, arrow_scale_fun = function(x) {     x }, labels = TRUE,
  parse_labels = FALSE, font = "Avenir", comp_fontsize = 1,
  arrow_fontsize = 1, filled = TRUE, alpha = 1, color_scaling = "RSE",
  color_cutoff = c(25, 50), arrow_color_manual = NULL,
  comp_color_manual = NULL, unscaled_color = NULL,
  unscaled_shape = "circle", scaled_shape = "square", flipped = FALSE,
  rank = NULL, title = NULL, clearance_mode = FALSE,
  pbpk_layout = FALSE, vein_comp_label = "venous",
  artery_comp_label = "arterial", save_qmd = FALSE, file_name = NULL,
  format = "svg", width = NULL, height = NULL, graph_attrs = NULL)
}
\arguments{
\item{qmd_info}{a \code{qmd_info} object generated with \code{import_qmd_info} or
\code{skeleton_qmd_info}}

\item{scaling}{logical if \code{TRUE} compartment size and colors will be scaled.
If \code{FALSE} standard model diagram will be created}

\item{comp_scale_fun}{a function to be used for compartment size scaling}

\item{arrow_scale_fun}{a function to be used for arrow width scaling}

\item{labels}{logical if \code{TRUE} labels are added to arrows}

\item{parse_labels}{logical if \code{TRUE} labels will be parsed to use subscript the rate constant name (KA will become K subscript A)
and the inter-compartmental clearance number (Q2 would become Q subscript 2)}

\item{font}{font name of the compartment label}

\item{comp_fontsize}{font size expansion factor}

\item{arrow_fontsize}{font size expansion factor}

\item{filled}{logical if \code{TRUE} compartment will be filled
If \code{FALSE} only the compartment edges will be drawn}

\item{alpha}{transparency factor}

\item{color_scaling}{can be 'iiv', 'rse', 'none' or 'pbpk'}

\item{color_cutoff}{numeric vector of length 2 setting the cutoff limits in color coding
for RSE (\%) or IIV (\%)}

\item{arrow_color_manual}{manually set color for each arrow}

\item{comp_color_manual}{manually set color for each compartment}

\item{unscaled_color}{color of the unscaled compartments}

\item{unscaled_shape}{shape of the unscaled compartments. Can be square, circle or diamond}

\item{scaled_shape}{shape of the scaled compartments. Can be square, circle or diamond}

\item{flipped}{logical if \code{TRUE} the layout will be flipped}

\item{rank}{integer vertor assigning a rank to each compartment. Can be used
to obtain a specific layout. The ranks must be greater or equal to 1.}

\item{title}{A title to be added to the graph}

\item{clearance_mode}{logical if \code{TRUE} clearances will be represented by triangles
and their surface area will be proportional to the volume cleared per unit of time}

\item{pbpk_layout}{logical if \code{TRUE} a PBPK layout will be applied}

\item{vein_comp_label}{label of the veinous compartment}

\item{artery_comp_label}{label of the arterial compartment}

\item{save_qmd}{save the QMD graph into a file (default \code{FALSE})}

\item{file_name}{name of the file to be created on the disk when save_qmd is \code{TRUE}}

\item{format}{file format when save_qmd is \code{TRUE}. Must be one of "png", "pdf", "svg" or "ps"}

\item{width}{width of the resulting graphic in pixels}

\item{height}{height of the resulting graphic in pixels}

\item{graph_attrs}{an optional data.frame of graph attribute statements that can
serve as defaults for the graph.}
}
\description{
Create and renders Quantitative Model Diagrams (QMD)
for NONMEM models
}
\examples{
\dontrun{
qmd(qmd_info, scaling = FALSE)
}
}
\seealso{
\code{\link{import_qmd_info}}, \code{\link{format_qmd_info}}
}
