% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nmtab.R
\name{read_nmtab}
\alias{read_nmtab}
\title{NONMEM output table import function}
\usage{
read_nmtab(file = NULL, skip = NULL, header = NULL,
  rm_duplicates = FALSE, nonmem_tab = TRUE)
}
\arguments{
\item{file}{full file name}

\item{skip}{number of lines to skip before reading data}

\item{header}{logical value indicating whether the file contains the names
of the variables as its first line}

\item{rm_duplicates}{logical value indicating whether duplicated columns should be removed}

\item{nonmem_tab}{logical value indicading to the function whether the file is a
table or a nonmem additional output file.}
}
\value{
A \code{data.frame}
}
\description{
Quickly import NONMEM output tables into R.
When both \code{skip} and \code{header} are \code{NULL},
\code{read_nmtab} will automatically detect the optimal
settings to import the tables. When more than one files are
provided for a same NONMEM run, they will be combined into
a single \code{data.frame}.
}
\examples{
\dontrun{
data <- read_nmtab(file = '../models/pk/sdtab101')
}
}
\seealso{
\code{\link{parse_nonmem_model}}, \code{\link{import_qmd_info}}
}
