% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_qmd.R
\name{save_qmd}
\alias{save_qmd}
\title{Save QMD to a file}
\usage{
save_qmd(qmd_graph = NULL, title = NULL, filename = "qmd_graph",
  width = NULL, height = NULL, device = "pdf")
}
\arguments{
\item{qmd_graph}{A qmd object generated by \code{render_graph}}

\item{title}{A title to be added to the graph}

\item{filename}{Name of the file to be created on the disk without the extension}

\item{width}{width of the output in pixels}

\item{height}{height of the output in pixels}

\item{device}{Device to use, can be "pdf", "ps", "jpeg", "png", or "svg"}
}
\description{
Save QMD to different file formats
}

