% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skeleton_qmd_info.R
\name{skeleton_qmd_info}
\alias{skeleton_qmd_info}
\title{Create skeleton for \code{qmd_info}}
\usage{
skeleton_qmd_info(help = TRUE)
}
\arguments{
\item{help}{logical if \code{TRUE} help will be added in the \code{qmd_info} skeleton.}
}
\value{
A list containing the individuals (\code{tvprm}) and population (\code{data}) parameters,
parameter uncertainty (\code{rse}) the nonmem ADVAN (\code{advan}) and TRANS (\code{trans})
as well as the parsed differencial equation (\code{des_info}).
}
\description{
Create skeleton for \code{qmd_info} as a manual alternative to \code{import_qmd_info}
}
\examples{
\dontrun{
qmd_info <- skeleton_qmd_info()
}
}
\seealso{
\code{\link{import_qmd_info}}, \code{\link{read_nmtab}}, \code{\link{parse_nommem_model}}
}

