% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/threecomp.R
\name{threecomp}
\alias{threecomp}
\title{Ciclosporin pharmacokinetic model parameters}
\format{A list of 3 levels: prm, rse and advan.}
\source{
S. Fanta et al. Developmental pharmacokinetics of ciclosporin –
a population pharmacokinetic study in paediatric renal transplant candidates.
British Journal of Clinical Pharmacology. 64:6, 772–784. (2007)
\url{http://www.ncbi.nlm.nih.gov/pubmed/17662086}
}
\description{
A dataset containing the pharmacokinetic typical parameters values and uncertainty
for ciclosporin in paediatric renal transplant candidates.
}
\details{
\itemize{
  \item prm. the model parameters typical values: KA (h-1), CL (L/h), V2 (L),
  Q2 (L/h), V3 (L), Q3 (L/h) and V4 (L)
  \item rse. the model parameters uncertainty as relative standard error (rse)
  \item advan. the nonmem subroutine ADVAN defining a 3-compartment model with
  first order absorption
}
}

