% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qmd.R
\name{create_qmd}
\alias{create_qmd}
\title{Creation and rendering of QMD graphs}
\usage{
create_qmd(comp = NULL, arrow = NULL, pbpk = NULL, graph_attrs = NULL,
  flipped = FALSE, title = NULL, save_qmd = FALSE, file_name = NULL,
  format = "svg", width = NULL, height = NULL)
}
\arguments{
\item{comp}{a data.frame created by \code{define_comp_layout}}

\item{arrow}{a data.frame created by \code{define_arrow_layout}}

\item{pbpk}{a list created by \code{define_pbpk_layout}}

\item{graph_attrs}{an optional data.frame of graph attribute statements that can
serve as defaults for the graph}

\item{flipped}{logical if \code{TRUE} the layout will be flipped}

\item{title}{a title to be added to the graph}

\item{save_qmd}{save the graph into a file (default \code{FALSE})}

\item{file_name}{name of the file to be created on the disk when save_qmd is \code{TRUE}}

\item{format}{file format when save_qmd is \code{TRUE}. Must be one of "png", "pdf", "svg" or "ps"}

\item{width}{width of the resulting graphic in pixels}

\item{height}{height of the resulting graphic in pixels}
}
\description{
Creates a QMD graph using data.frames for comp and arrow.
}
\examples{
\dontrun{
define_graph(comp  = define_comp_layout(examples$onecomp),
             arrow = define_arrow_layout(examples$onecomp))
}
}
