% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/define_arrow_layout.R
\name{define_arrow_layout}
\alias{define_arrow_layout}
\title{Create arrow connections between compartments}
\usage{
define_arrow_layout(qmd_info = NULL, scaling = TRUE,
  scale.fun = function(x) {     x^(1/3) }, font = "Avenir",
  clearance = TRUE, arrow.fontsize = 12, ...)
}
\arguments{
\item{qmd_info}{a \code{qmd_info} object generated with \code{import_qmd_info} or
\code{skeleton_qmd_info}}

\item{scaling}{logical, if \code{TRUE} arrow width and colors will be scaled.
If \code{FALSE} standard model diagram will be created}

\item{scale.fun}{a function to be used for arrow width scaling}

\item{font}{font name of arrow labels}

\item{clearance}{logical, if \code{TRUE} arrows will be scaled to their available
clearance value. If \code{FALSE} arrows will be scaled to their available
rate constant value}

\item{arrow.fontsize}{font size of the arrow labels in point size}
}
\value{
A \code{data.frame}
}
\description{
Combine information for each arrow (origin, destination, label,
witdh, color, scaling and more) into a dataframe
}
\details{
The default \code{scaling.fun} argument is set to \code{cubic}, to match
 each the method used on volumes. Arrow will be sized to the cubic root of the
 clearance or rate constant value.
}
\examples{
\dontrun{
qmd_info <- import_qmd_info(dir = '../models/pk/', runno = '001')
arrows   <- define_arrow_layout(qmd_info)
}
}
\seealso{
\code{\link{define_comp_layout}}, \code{\link{import_qmd_info}}, \code{\link{qmd}}
}

