% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_comp_layout.R
\name{define_comp_layout}
\alias{define_comp_layout}
\title{Create compartment layout}
\usage{
define_comp_layout(qmd_info = NULL, scaling = TRUE,
  comp_scale_fun = function(x) {     sqrt(x) }, color_scaling = "RSE",
  color_cutoff = c(25, 50), filled = TRUE, alpha = 1,
  comp_color_manual = NULL, unscaled_color = NULL,
  unscaled_shape = "circle", scaled_shape = "square", font = "Avenir",
  comp_fontsize = 1)
}
\arguments{
\item{qmd_info}{a \code{qmd_info} object generated with \code{import_qmd_info} or
\code{skeleton_qmd_info}}

\item{scaling}{logical if \code{TRUE} compartment size and colors will be scaled.
If \code{FALSE} standard model diagram will be created}

\item{comp_scale_fun}{a function to be used for compartment size scaling}

\item{color_scaling}{can be 'iiv', 'rse', 'none' or 'pbpk'}

\item{color_cutoff}{numeric vector of length 2 setting the cutoff limits in color coding
for RSE (\%) or IIV (\%)}

\item{filled}{logical if \code{TRUE} compartment will be filled
If \code{FALSE} only the compartment edges will be drawn}

\item{alpha}{transparency factor}

\item{comp_color_manual}{manually set color for each compartment}

\item{unscaled_color}{color of the unscaled compartments}

\item{unscaled_shape}{shape of the unscaled compartments. Can be square, circle or diamond}

\item{scaled_shape}{shape of the scaled compartments. Can be square, circle or diamond}

\item{font}{font name of the compartment label}

\item{comp_fontsize}{font size expansion factor}
}
\value{
A \code{data.frame}
}
\description{
Combine information for each compartment's rank, label, size,
color and scaling into a \code{data.frame}.
}
\details{
The default \code{scaling_fun} argument is set to be a surface area
of the compartment.
}
\examples{
\dontrun{
qmd_info <- import_qmd_info(dir = '../models/pk/', runno = '001')
comp     <- define_comp_layout(qmd_info)
}
}
\seealso{
\code{\link{define_arrow_layout}}, \code{\link{import_qmd_info}}, \code{\link{qmd}}
}

