% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/define_graph.R
\name{define_graph}
\alias{define_graph}
\title{DiagrammeR functions}
\usage{
define_graph(nodes_df = NULL, edges_df = NULL, graph_attrs = NULL,
  node_attrs = NULL, edge_attrs = NULL, directed = TRUE,
  graph_name = NULL, graph_time = NULL, graph_tz = NULL)
}
\description{
Edited from DiagrammeR to include ranking option but remain compatible
}

