% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/format_qmd_info.R
\name{format_qmd_info}
\alias{format_qmd_info}
\title{Parameter translator}
\usage{
format_qmd_info(qmd_info = NULL)
}
\arguments{
\item{qmd_info}{a \code{qmd_info} object generated with \code{import_qmd_info}
or \code{skeleton_qmd_info}}
}
\value{
A \code{data.frame} \code{qmd_info} object
}
\description{
Translate clearances into rate constant or rate constant into
clearances for the built-in nonmem model library (i.e. ADVAN 1-4 and 11-12)
}
\examples{
\dontrun{
qmd_info  <- import_qmd_info(dir = '../models/pk/', runno = '001')
qmd_info_formatted <- format_qmd_info(qmd_info)
}
}
\seealso{
\code{\link{import_qmd_info}}, \code{\link{skeleton_qmd_info}}, \code{\link{qmd}}
}

