% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_qmd_info.R
\name{import_qmd_info}
\alias{import_qmd_info}
\title{Import model parameters}
\usage{
import_qmd_info(dir = NULL, prefix = "run", runno = NULL, ext = ".mod",
  file = NULL, interactive = TRUE, verbose = FALSE)
}
\arguments{
\item{dir}{location of the model files}

\item{prefix}{prefix of the model file name}

\item{runno}{run number to be evaluated}

\item{ext}{model file extention}

\item{file}{full file name as an alternative to \code{dir}, \code{prefix},
\code{runno} and \code{ext}}

\item{interactive}{if \code{TRUE} parameter labels can not be matched with their thetas, omega
and sigma, the user will be asked to manually provide labels. If \code{FALSE} an error will be
returned instead}

\item{verbose}{if \code{TRUE} messages are printed to the console}
}
\value{
A list containing the individuals (\code{tvprm}) and population (\code{data}) parameters,
parameter uncertainty (\code{rse}) the nonmem ADVAN (\code{advan}) and TRANS (\code{trans})
as well as the parsed differencial equation (\code{des_info}).
}
\description{
Import model parameters from NONMEM run
}
\examples{
\dontrun{
qmd_info <- import_qmd_info(dir = '../models/pk/', runno = '001')
}
}
\seealso{
\code{\link{format_qmd_info}}, \code{\link{qmd}}
}

