% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qmd_info.R
\name{import_qmd_info}
\alias{import_qmd_info}
\title{Import model parameters}
\usage{
import_qmd_info(dir = NULL, prefix = "run", runno = NULL, ext = ".mod",
  file = NULL, interactive = TRUE, verbose = FALSE)
}
\arguments{
\item{dir}{location of the model files}

\item{prefix}{prefix of the model file name}

\item{runno}{run number to be evaluated}

\item{ext}{model file extention}

\item{file}{full file name as an alternative to \code{dir}, \code{prefix},
\code{runno} and \code{ext}}

\item{interactive}{if \code{TRUE} parameter labels can not be associated with their thetas, omega
and sigma, the user will be asked to manually provide labels. If \code{FALSE} an error will be
returned instead}

\item{verbose}{if \code{TRUE} messages are printed to the console}
}
\value{
A list containing the fixed effect (\code{theta}), random effect variance (\code{omega})
typical values along with their uncertainty, the indivudual parameters (\code{data})
the nonmem ADVAN (\code{advan}), the model differential equations (\code{des}),
the parsed compartment information (\code{parsed_comp}),
and the parsed arrow information (\code{parsed_arrow}).
}
\description{
Import model parameters from NONMEM run
}
\examples{
\dontrun{
qmd_info <- import_qmd_info(dir = '../models/pk/', runno = '001')
}
}
\seealso{
\code{\link{format_qmd_info}}, \code{\link{qmd}}
}

