% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelviz.R
\name{modelviz}
\alias{modelviz}
\title{Create and render QMD}
\usage{
modelviz(data, horizontal = TRUE, ...)
}
\arguments{
\item{data}{a \code{list} containing the parameters, their RSE and the
nonmem $SUB ADVAN}

\item{horizontal}{logical if \code{TRUE} the layout will be horizontal}
}
\value{
A graphic object
}
\description{
Create and renders Quantitative Model Diagrams (QMD)
}
\examples{
\dontrun{
modelviz(parameters, horizontal = FALSE)
}
}
\seealso{
\code{\link{prm_import}}, \code{\link{prm_format}}
}

