% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qmd.R
\name{qmd}
\alias{qmd}
\title{Create and render QMD}
\usage{
qmd(qmd_info = NULL, horizontal = TRUE, shiny = FALSE, ...)
}
\arguments{
\item{qmd_info}{a \code{list} containing the parameters, their RSE and the
nonmem subroutine ADVAN value}

\item{horizontal}{logical if \code{TRUE} the layout will be horizontal}

\item{shiny}{logical if \code{TRUE} output will be formated for shiny output}

\item{...}{other arguments passed to \code{\link{define_comp_layout}}
and \code{\link{define_arrow_layout}} see details below for
a list of the available options}
}
\value{
A graphic object
}
\description{
Create and renders Quantitative Model Diagrams (QMD)
for NONMEM models
}
\details{
\itemize{
 \item{scaling}
 \item{scale.fun}
 \item{font}
 \item{comp.fontsize}
 \item{arrow.fontsize}
 \item{box.ratio}
 \item{filled}
 \item{clearance}
}
}
\examples{
\dontrun{
qmd(qmd_info, horizontal = FALSE)
}
}
\seealso{
\code{\link{import_qmd_info}}, \code{\link{format_qmd_info}}
}

