% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd.R
\name{qmd}
\alias{qmd}
\title{Create and render QMD}
\usage{
qmd(qmd_info = NULL, scaling = TRUE, comp_scale_fun = function(x) {    
  sqrt(x) }, arrow_scale_fun = function(x) {     x }, labels = TRUE,
  parse_labels = FALSE, font = "Avenir", comp_fontsize = 1,
  arrow_fontsize = 1, filled = TRUE, alpha = 1, color_scaling = "RSE",
  color_cutoff = c(25, 50), arrow_color_manual = NULL,
  comp_color_manual = NULL, unscaled_color = NULL,
  unscaled_shape = "circle", scaled_shape = "square", flipped = FALSE,
  rank = NULL, clearance_mode = FALSE, pbpk_layout = FALSE,
  vein_comp_label = "venous", artery_comp_label = "arterial",
  shiny = FALSE, output = "graph", filename = "qmd_graph", title = NULL,
  width = NULL, height = NULL, gv_options = NULL)
}
\arguments{
\item{qmd_info}{a \code{list} containing the parameters, their RSE and the
nonmem subroutine ADVAN value}

\item{scaling}{logical, if \code{TRUE} arrow width and colors will be scaled.
If \code{FALSE} standard model diagram will be created}

\item{comp_scale_fun}{a function to be used for compartment size scaling}

\item{arrow_scale_fun}{a function to be used for arrow width scaling}

\item{labels}{logical if \code{TRUE} labels are added to arrows}

\item{parse_labels}{logical if \code{TRUE} arrow labels will be parsed to use subscript the rate
constant name (\deqn{KA} will become \deqn{K_A}) and the inter-compartmental clearance number
(\deqn{Q2} would become \deqn{Q_2})}

\item{font}{font name of the compartment label}

\item{comp_fontsize}{font size expansion factor}

\item{arrow_fontsize}{font size expansion factor}

\item{filled}{logical if \code{TRUE} compartment will be filled
If \code{FALSE} only the compartment edges will be drawn}

\item{alpha}{transparency factor}

\item{color_scaling}{can be 'iiv', 'rse', 'none' or 'pbpk'}

\item{color_cutoff}{numeric vector of length 2 setting the cutoff limits in color coding
for RSE (\%) or IIV (\%)}

\item{arrow_color_manual}{manually set color for each arrow}

\item{comp_color_manual}{manually set color for each compartment}

\item{unscaled_color}{color of the unscaled compartments}

\item{unscaled_shape}{shape of the unscaled compartments. Can be square, circle or diamond}

\item{scaled_shape}{shape of the scaled compartments. Can be square, circle or diamond}

\item{flipped}{logical if \code{TRUE} the layout will be flipped}

\item{rank}{integer vertor assigning a rank for each compartment. Can be used
to obtain a specific layout}

\item{clearance_mode}{logical if \code{TRUE} clearances will be represented by triangles
and their surface area will be proportional to the volume cleared per unit of time}

\item{pbpk_layout}{logical if \code{TRUE} a PBPK layout will be applied}

\item{vein_comp_label}{label of the veinous compartment}

\item{artery_comp_label}{label of the arterial compartment}

\item{shiny}{logical if \code{TRUE} output will be formated for shiny output}

\item{output}{format of the output to be returned by the qmd function ('graph', 'vivagraph',
'visNetwork'), or an output format to save the graph ('pdf', 'ps', 'jpeg',
'png', or 'svg')}

\item{filename}{Name of the file to be created on the disk (without extension) when output is one
of ('pdf', 'ps', 'jpeg', 'png', or 'svg')}

\item{title}{A title to be added to the graph}

\item{width}{width of the resulting graphic in pixels}

\item{height}{height of the resulting graphic in pixels}

\item{gv_options}{vector of options to be passed to Graphviz}
}
\value{
A graphic object
}
\description{
Create and renders Quantitative Model Diagrams (QMD)
for NONMEM models
}
\examples{
\dontrun{
qmd(qmd_info, scaling = FALSE)
}
}
\seealso{
\code{\link{import_qmd_info}}, \code{\link{format_qmd_info}}
}

