% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qmd.R
\name{qmd}
\alias{qmd}
\title{Create and render QMD}
\usage{
qmd(qmd_info = NULL, flipped = FALSE, rank = NULL, pbpk_layout = FALSE,
  color_scaling = NULL, shiny = FALSE, output = "graph", width = NULL,
  height = NULL, ...)
}
\arguments{
\item{qmd_info}{a \code{list} containing the parameters, their RSE and the
nonmem subroutine ADVAN value}

\item{flipped}{logical if \code{TRUE} the layout will be flipped}

\item{rank}{integer vertor assigning a rank for each compartment. Can be used
to obtain a specific layout}

\item{pbpk_layout}{logical if \code{TRUE} a PBPK layout will be applied}

\item{shiny}{logical if \code{TRUE} output will be formated for shiny output}

\item{output}{format of the output to be returned by qmd ('graph', 'SVG', 'DOT' or 'vivagraph')}

\item{width}{width of the resulting graphic in pixels}

\item{height}{height of the resulting graphic in pixels}

\item{...}{other arguments passed to \code{\link{define_comp_layout}},
\code{\link{define_arrow_layout}} and \code{\link{define_pbpk_layout}} see details below for
a list of the available options}
}
\value{
A graphic object
}
\description{
Create and renders Quantitative Model Diagrams (QMD)
for NONMEM models
}
\details{
\itemize{
 \item{scaling}
 \item{comp_scale_fun}
 \item{arrow_scale_fun}
 \item{color_scaling}
 \item{color_cutoff}
 \item{pbpk_color}
 \item{filled}
 \item{labels}
 \item{alpha}
 \item{font}
 \item{comp_fontsize}
 \item{arrow_fontsize}
 \item{clearance_mode}
 \item{vein_comp_label}
 \item{artery_comp_label}
}
}
\examples{
\dontrun{
qmd(qmd_info, flipped = FALSE)
}
}
\seealso{
\code{\link{import_qmd_info}}, \code{\link{format_qmd_info}}
}

