% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skeleton_qmd_info.R
\name{skeleton_qmd_info}
\alias{skeleton_qmd_info}
\title{Create skeleton for \code{qmd_info}}
\usage{
skeleton_qmd_info(help = TRUE)
}
\arguments{
\item{help}{logical if \code{TRUE} help will be added in the \code{qmd_info} skeleton.}
}
\value{
A list containing the fixed effect (\code{theta}), random effect variance (\code{omega})
typical values along with their uncertainty, the indivudual parameters (\code{data})
the nonmem ADVAN (\code{advan}), the parsed compartment information (\code{parsed_comp}),
and the parsed arrow information (\code{parsed_arrow}).
}
\description{
Create skeleton for \code{qmd_info} as a manual alternative to \code{import_qmd_info}
}
\examples{
\dontrun{
qmd_info <- skeleton_qmd_info()
}
}
\seealso{
\code{\link{import_qmd_info}}, \code{\link{read_nmtab}}, \code{\link{parse_nonmem_model}}
}

